refactoring  TabBar
